Move centroiding functions to separate file
